[ptf_runner] rename ptf.py to ptf_runner.py to avoid conflict with ptf framework
